chore: 머지 안정화 및 infinite recursion 오류 수정
@JsonManagedReference
@JsonBackReference

+++ src/main/java/com/example/pens/domain/request/GroupDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class GroupDTO {
-    private Long groupId;
+    private int groupId;
     private String groupName;
     private String groupAdmin;
 }
